+++ src/materials/Glass.h
@@ -6,7 +6,7 @@ public:
   Glass(
     Vector3D shine=Vector3D(1,1,1),
     Vector3D transparency=Vector3D(1,1,1),
-    double indexOfRefraction=1.5,
-    double roughness=0
+    float indexOfRefraction=1.5,
+    float roughness=0
   ) : Material(shine, transparency, indexOfRefraction, roughness) {};
 };
